Break out more submodules for array-specific stuff and plotting
